[API] [WEB] - Refactor Organization routes to use tenant with super user
#1669, #1672

+++ marfim-api/src/main/java/org/ipdec/marfim/api/service/OrganizationService.java
@@ -6,10 +6,12 @@ import org.ipdec.marfim.api.exception.OrganizationException;
 import org.ipdec.marfim.api.exception.type.OrganizationExceptionsEnum;
 import org.ipdec.marfim.api.model.Organization;
 import org.ipdec.marfim.api.repository.OrganizationRepository;
+import org.ipdec.marfim.security.IPrincipalTokenAttributes;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 
@@ -19,9 +21,14 @@ public class OrganizationService {
 
     private final OrganizationRepository organizationRepository;
 
+    private final IPrincipalTokenAttributes principal;
 
-    public Organization findById(Long id) {
-        return organizationRepository.findById(id).orElseThrow(()->{
+
+    public Organization findById(Long organizationId, Long tenantId) {
+        if(!principal.getUser().isSuper() && !Objects.equals(tenantId, organizationId)){
+            throw new OrganizationException(OrganizationExceptionsEnum.FORBIDDEN_ORGANIZATION_ID_DOES_NOT_MATCH);
+        }
+        return organizationRepository.findById(organizationId).orElseThrow(()->{
             return new OrganizationException(OrganizationExceptionsEnum.NOT_FOUND);
         });
     }
@@ -32,7 +39,7 @@ public class OrganizationService {
 
     public List<Organization> findAll(Long organizationId) {
         List<Organization> allOrganizations = findAll();
-        if(organizationId==null) {
+        if(principal.getUser().isSuper()) {
             return allOrganizations;
         }
 
@@ -43,6 +50,9 @@ public class OrganizationService {
     }
 
     public Organization create(CreateOrganizationDTO organizationDTO){
+        if(!principal.getUser().isSuper()){
+            throw new OrganizationException(OrganizationExceptionsEnum.FORBIDDEN_CREATE_AND_DELETE_BY_NON_SUPER_USER);
+        }
         Organization organization = organizationDTO.parseToOrganization(new Organization());
 
         //TODO: check valid cnpj
@@ -59,8 +69,8 @@ public class OrganizationService {
         return organization;
     }
 
-    public Organization update(Long id, CreateOrganizationDTO organizationDTO){
-        Organization organization = organizationDTO.parseToOrganization(findById(id));
+    public Organization update(Long id, CreateOrganizationDTO organizationDTO, Long tenantId){
+        Organization organization = organizationDTO.parseToOrganization(findById(id, tenantId));
 
         //TODO: check valid cnpj
 
@@ -77,8 +87,11 @@ public class OrganizationService {
         return organization;
     }
 
-    public void delete(Long id) {
-        findById(id);
+    public void delete(Long id, Long tenantId) {
+        if(!principal.getUser().isSuper()){
+            throw new OrganizationException(OrganizationExceptionsEnum.FORBIDDEN_CREATE_AND_DELETE_BY_NON_SUPER_USER);
+        }
+        findById(id, tenantId);
         organizationRepository.deleteById(id);
     }
